chore: 셀 오픈 기능 단위 테스트, CellTest.kt 로 이동
